requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
dacite==1.6.0
ftfy==6.1.1
idna==3.3
joblib==1.1.0
langcodes==3.3.0
msgpack==1.0.3
nltk==3.7
pyenchant==3.2.2
PyYAML==6.0
regex==2022.1.18
requests==2.27.1
tqdm==4.62.3
urllib3==1.26.8
wcwidth==0.2.5
wordfreq==2.5.1
